% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDACdecon.R
\name{PDACdecon}
\alias{PDACdecon}
\title{Create Profile Matrix for SpatialDecon from CellDive Data.}
\usage{
PDAC_Decon(
  x,
  bulk_counts,
  celltypes,
  protein_to_gene = TRUE,
  correlation = 0.4,
  weight_factor = 0.2,
  verbose = TRUE,
  org.db,
  match_table,
  org.db.from = "ALIAS",
  org.db.to = "SYMBOL"
)
}
\arguments{
\item{x}{celldive average expression dataset, a data.frame or a matrix}

\item{bulk_counts}{bulk RNA expression dataset, a data.frame or a matrix}

\item{celltypes}{a data.frame which must have two columns: gene, cell}

\item{protein_to_gene}{logical value, if TURE rownames of celldive will be used as protein, and will be mapped to gene id}

\item{correlation}{most similar gene cutoff}

\item{weight_factor}{weight value of a gene when it was a celltype marker}

\item{verbose}{if TRUE, then more message will be print}

\item{org.db}{org.db object, only use when protein_to_gene is TURE}

\item{match_table}{a data.frame which must have two columns: FROM, TO, only use when protein_to_gene is TURE}

\item{org.db.from}{keytypes which annotationdbi used, only use when protein_to_gene is TURE}

\item{org.db.to}{keytypes which annotationdbi used, only use when protein_to_gene is TURE}
}
\description{
This function generate a Profile Matrix for SpatialDecon from CellDive Data.
}
\examples{
# CellDive2SpatialDecon()

}
