% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{protein2gene}
\alias{protein2gene}
\title{Transform protein name to gene symbol}
\usage{
protein2gene(
  x,
  org.db,
  match_table,
  org.db.from = "ALIAS",
  org.db.to = "SYMBOL"
)
}
\arguments{
\item{x}{character vector, protein id used to be mapped to gene}

\item{org.db}{org.db object}

\item{match_table}{a data.frame which must have two columns: FROM, TO}

\item{org.db.from}{keytypes which annotationdbi used}

\item{org.db.to}{keytypes which annotationdbi used}
}
\description{
Transform protein name to gene symbol
}
\examples{
# protein2gene()

}
